Task 3 - Homing Missile Completed
I'm not sure what this stuff about the BillingMode and Resources is but google says its okay and git will fix it so I'm hoping this will be okay? If the download is broken for whatever reason feel free to contact me and I can send the proper build, but Google says it should be fine.